move out init_spectral_library
